Make JS files Modular
