+++ index.html
@@ -51,6 +51,6 @@
         <h3>copyright -</h3>
       </nav>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
